:sparkles: Added `dataCollectionSetUp`
