+++ base/src/main/java/com/arnyminerz/escalaralcoiaicomtat/activity/NewMainActivity.kt
@@ -55,6 +55,7 @@ import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import coil.compose.rememberImagePainter
 import com.arnyminerz.escalaralcoiaicomtat.R
+import com.arnyminerz.escalaralcoiaicomtat.core.firebase.dataCollectionSetUp
 import com.arnyminerz.escalaralcoiaicomtat.core.ui.element.Backdrop
 import com.arnyminerz.escalaralcoiaicomtat.core.utils.toast
 import com.arnyminerz.escalaralcoiaicomtat.ui.theme.EscalarAlcoiaIComtatTheme
@@ -64,6 +65,8 @@ import com.arnyminerz.escalaralcoiaicomtat.ui.theme.EscalarAlcoiaIComtatTheme
 class NewMainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+
+        dataCollectionSetUp()
 
         setContent {
             EscalarAlcoiaIComtatTheme {
